+++ src-docs/src/views/tables/actions/actions_section.js
@@ -32,11 +32,13 @@ export const section = {
       </p>
       <ul>
         <li>
-          There can only be a single action tool visible per row. When more than one action is defined,
-          they will collapse under a single popover represented by the gear icon.
+          There can only be up to 2 actions visible per row. When more than two actions are defined,
+          the first 2 <code>isPrimary</code> actions will stay visible, an ellipses icon button will hold all actions
+          in a single popover.
         </li>
         <li>
-          Actions are only visible when the user hovers over the row with the mouse.
+          Actions are change opacity when user hovers over the row with the mouse. When more than 2 actions are supplied,
+          only the ellipses icon button stays visible at all times.
         </li>
       </ul>
     </div>