+++ src/components/search_bar/filters/field_value_selection_filter.test.js
@@ -65,6 +65,44 @@ describe('FieldValueSelectionFilter', () => {
 
   });
 
+  test('render - fields in options', () => {
+
+    const props = {
+      ...requiredProps,
+      index: 0,
+      onChange: () => {},
+      query: Query.parse(''),
+      config: {
+        type: 'field_value_selection',
+        name: 'Tag',
+        options: [
+          {
+            field: 'tag',
+            value: 'feature'
+          },
+          {
+            field: 'tag_2',
+            value: 'test',
+            name: 'Text'
+          },
+          {
+            field: 'tag_3',
+            value: 'bug',
+            name: 'Bug',
+            view: <div>bug</div>
+          }
+        ]
+      }
+    };
+
+    const component = shallow(
+      <FieldValueSelectionFilter {...props} />
+    );
+
+    expect(component).toMatchSnapshot();
+
+  });
+
   test('render - all configurations', () => {
 
     const props = {
@@ -110,6 +148,154 @@ describe('FieldValueSelectionFilter', () => {
         noOptionsMessage: 'oops...',
         searchThreshold: 5,
         options: () => {}
+      }
+    };
+
+    const component = shallow(
+      <FieldValueSelectionFilter {...props} />
+    );
+
+    expect(component).toMatchSnapshot();
+
+  });
+
+  test('inactive - field is global', () => {
+
+    const props = {
+      ...requiredProps,
+      index: 0,
+      onChange: () => {},
+      query: Query.parse(''),
+      config: {
+        type: 'field_value_selection',
+        field: 'tag',
+        name: 'Tag',
+        options: [
+          {
+            value: 'feature'
+          },
+          {
+            value: 'test',
+            name: 'Text'
+          },
+          {
+            value: 'bug',
+            name: 'Bug',
+            view: <div>bug</div>
+          }
+        ]
+      }
+    };
+
+    const component = shallow(
+      <FieldValueSelectionFilter {...props} />
+    );
+
+    expect(component).toMatchSnapshot();
+
+  });
+
+  test('active - field is global', () => {
+
+    const props = {
+      ...requiredProps,
+      index: 0,
+      onChange: () => {},
+      query: Query.parse('tag:bug'),
+      config: {
+        type: 'field_value_selection',
+        field: 'tag',
+        name: 'Tag',
+        options: [
+          {
+            value: 'feature'
+          },
+          {
+            value: 'test',
+            name: 'Text'
+          },
+          {
+            value: 'bug',
+            name: 'Bug',
+            view: <div>bug</div>
+          }
+        ]
+      }
+    };
+
+    const component = shallow(
+      <FieldValueSelectionFilter {...props} />
+    );
+
+    expect(component).toMatchSnapshot();
+
+  });
+
+  test('inactive - fields in options', () => {
+
+    const props = {
+      ...requiredProps,
+      index: 0,
+      onChange: () => {},
+      query: Query.parse(''),
+      config: {
+        type: 'field_value_selection',
+        name: 'Tag',
+        options: [
+          {
+            field: 'tag',
+            value: 'feature'
+          },
+          {
+            field: 'tag_2',
+            value: 'test',
+            name: 'Text'
+          },
+          {
+            field: 'tag_3',
+            value: 'bug',
+            name: 'Bug',
+            view: <div>bug</div>
+          }
+        ]
+      }
+    };
+
+    const component = shallow(
+      <FieldValueSelectionFilter {...props} />
+    );
+
+    expect(component).toMatchSnapshot();
+
+  });
+
+  test('active - fields in options', () => {
+
+    const props = {
+      ...requiredProps,
+      index: 0,
+      onChange: () => {},
+      query: Query.parse('tag_3:bug'),
+      config: {
+        type: 'field_value_selection',
+        name: 'Tag',
+        options: [
+          {
+            field: 'tag',
+            value: 'feature'
+          },
+          {
+            field: 'tag_2',
+            value: 'test',
+            name: 'Text'
+          },
+          {
+            field: 'tag_3',
+            value: 'bug',
+            name: 'Bug',
+            view: <div>bug</div>
+          }
+        ]
       }
     };