+++ src/components/search_bar/query/default_syntax.js
@@ -96,7 +96,7 @@ identifier
 
 identifierChar
   = alnum
-  / [-]
+  / [-_]
   / escapedChar
   
 fieldRangeValue
@@ -115,11 +115,11 @@ containsOrValues
   )* space? ")" { return [ head, ...tail ]; }
   
 rangeValue
-  = number
+  = numberWord
   / date
 
 containsValue
-  = number
+  = numberWord
   / date
   / booleanWord
   / word
@@ -159,6 +159,11 @@ boolean
 
 number
  = [\\-]?[0-9]+("."[0-9]+)* { return Exp.number(text(), location()); }
+
+// only match numbers followed by whitespace or end of input 
+numberWord
+ = num:number &space { return num; }
+ / num:number !. { return num; }
 
 date
  = "'" expression:((!"'" .)+ { return text(); }) "'" {
@@ -218,14 +223,20 @@ const resolveFieldValue = (field, valueExpression, ctx) => {
   if (isArray(valueExpression)) {
     return valueExpression.map(exp => resolveFieldValue(field, exp, ctx));
   }
-  const { type, expression, location } = valueExpression;
+  const { location } = valueExpression;
+  let { type, expression } = valueExpression;
   if (schema && !schema.fields[field] && schema.strict) {
     error(`Unknown field \`${field}\``, location);
   }
   const schemaField = schema && schema.fields[field];
   if (schemaField && schemaField.type !== type && schema.strict) {
-    const valueDesc = schemaField.valueDescription || `a ${schemaField.type} value`;
-    error(`Expected ${valueDesc} for field \`${field}\`, but found \`${expression}\``, location);
+    if (schemaField.type === 'string') {
+      expression = valueExpression.expression = expression.toString();
+      type = valueExpression.type = 'string';
+    } else {
+      const valueDesc = schemaField.valueDescription || `a ${schemaField.type} value`;
+      error(`Expected ${valueDesc} for field \`${field}\`, but found \`${expression}\``, location);
+    }
   }
   switch(type) {